+++ packages/schemas/src/candle/responses.ts
