+++ packages/schemas/src/index.ts
@@ -1,7 +1,10 @@
 export * from './auth';
-export * from './cake';
-export * from './letter';
-export * from './candle';
+export * from './cake/requests';
+export * from './cake/responses';
+export * from './letter/requests';
+export * from './letter/responses';
+export * from './candle/requests';
+export * from './candle/responses';
 
 export * from './user/schema';
 export * from './user/requests';
