+++ packages/schemas/src/user/responses.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const getUserYearRes = z.object({
+  year: z.number(),
+});
+export type getUserYearRes = z.infer<typeof getUserYearRes>;
